pyproject.toml
[project]
name = "live-checker"
version = "0.1.0"
description = "Сервис для проверки сервера на ответ"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram==3.27.0",
    "alembic>=1.18.4",
    "arq>=0.25.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.0",
    "gunicorn>=25.3.0",
    "httpx>=0.28.1",
    "pydantic>=2.4.1,<2.13",
    "pydantic-settings>=2.14.0",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.45.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
